+++ src/Trivial.Ui.GeekQuiz/Options/GeneralOptions.cs
@@ -120,5 +120,15 @@ namespace Trivial.Ui.GeekQuiz.Options
         [DisplayName(CommonConstants.SuppressClosingWithoutSubmitingAnswerWarningOptionLabel)]
         [Description(CommonConstants.SuppressClosingWithoutSubmitingAnswerWarningOptionDetailedDescription)]
         public bool SuppressClosingWithoutSubmitingAnswerWarning { get; set; } = false;
+
+        [Category(CommonConstants.CategorySubLevel)]
+        [DisplayName(CommonConstants.ShowTriviaUponClosingSolutionOptionLabel)]
+        [Description(CommonConstants.ShowTriviaUponClosingSolutionOptionDetailedDescription)]
+        public bool ShowTriviaUponClosingSolution { get; set; } = false;
+
+        [Category(CommonConstants.CategorySubLevel)]
+        [DisplayName(CommonConstants.ShowTriviaUponOpeningSolutionOptionLabel)]
+        [Description(CommonConstants.ShowTriviaUponOpeningSolutionOptionDetailedDescription)]
+        public bool ShowTriviaUponOpeningSolution { get; set; } = true;
     }
 }